+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ## Upcoming Release
 - Texture Node now uses the smallest of dimensions instead of width for calculating input resolution. This isn't considered breaking as this node is meant to be used with heightmaps or splatmaps, i.e. textures that are usually square
+- Added check that compute shaders and random write textures are actually supported
 
 ## v0.4.0
 - Wrote some compute shader replacements for regular shaders